Fix buffer resize and GpuLocked issues
Buffers would be resized to 64 times their size because of omitted
division by vertex struct length.

It was hard to use the lib in a subpass without blocking on the previous
frame to end, because the previous frame is using the buffers.
Unwrapping the write lock would crash the executable. Handle the case
where the buffers are in use by allocating new buffers of the same size.